// app/constants/type.ts
interface Performance {
	id: string
	title: string
	mainPhoto: string
	performanceType: string
	constructionStartDate: string
	constructionEndDate: string
	clientName: string
	contractAmount: string
	createdAt: string
	primeContractor: string
	subTable: string
	subPhoto1?: string
	subPhoto2?: string
	subPhoto3?: string
	subPhoto4?: string
	subPhoto5?: string
}

export type { Performance }